requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==18.2.0
backcall==0.1.0
bcrypt==3.1.6
billiard==3.3.0.23
bleach==3.1.0
captcha==0.3
celery==3.1.26.post2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
decorator==4.3.2
defusedxml==0.5.0
Django==1.11.20
django-celery==3.2.1
django-ckeditor==5.4.0
django-js-asset==1.2.2
django-redis==4.10.0
easy-thumbnails==2.6
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.7
Flask==1.0.2
Flask-Cors==3.0.6
idna==2.8
IPy==1.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itchat==1.3.10
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==3.0.37
lxml==4.4.1
mailer==0.8.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
openpyxl==2.6.1
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.3.4
pickleshare==0.7.5
Pillow==6.0.0
pipenv==2018.11.26
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
PyMySQL==0.9.3
PyNaCl==1.3.0
PyPDF2==1.26.0
pypng==0.0.19
PyQRCode==1.2.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pywinpty==0.5.5
pyzmq==18.0.0
qtconsole==4.4.3
redis==3.2.1
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
xlrd==1.2.0
XlsxWriter==1.1.5
